requirements.txt
ccxt
pandas
python-dotenv
ta-lib
plotly
MetaTrader5
xlsxwriter
